+++ libs/rxjs/array/src/utils/array-set.ts
@@ -1,0 +1,16 @@
+/**
+ * @packageDocumentation
+ * @module Array
+ */
+
+/**
+ * Returns if the input is an array or
+ * @private
+ * @internal
+ * @param input
+ */
+export function isArrayOrSet(input: unknown): input is ArrayLike<unknown> {
+  if (Array.isArray(input)) {
+    return true;
+  } else return input instanceof Set;
+}